Update firebird*.c: add casts for MSVC 64-bit warnings

+++ source/firebird.c
@@ -134,7 +134,7 @@ static void fb_log_status(PFB_SESSION session, const char *from)
     // util_output("%s%s", s, nl);
   }
 
-  session->errorcode = session->status[1];
+  session->errorcode = (long)session->status[1];
   HB_SYMBOL_UNUSED(from);
 
   if (session->transac) {
@@ -185,20 +185,20 @@ HB_FUNC(FBCONNECT)
   i = 0;
   dpb[i++] = isc_dpb_version1;
   dpb[i++] = isc_dpb_user_name;
-  len = strlen(user);
+  len = (int)strlen(user);
   dpb[i++] = (char)len;
   memcpy(&(dpb[i]), user, len);
   i += len;
 
   dpb[i++] = isc_dpb_password;
-  len = strlen(passwd);
+  len = (int)strlen(passwd);
   dpb[i++] = (char)len;
   memcpy(&(dpb[i]), passwd, len);
   i += len;
 
   if (charset != SR_NULLPTR) {
     dpb[i++] = isc_dpb_lc_ctype;
-    len = strlen(charset);
+    len = (int)strlen(charset);
     dpb[i++] = (char)len;
     memcpy(&(dpb[i]), charset, len);
     i += len;
@@ -1167,7 +1167,7 @@ HB_FUNC(FBLINEPROCESSED)
   HB_SIZE lLen, lDec;
 
   if (session) {
-    cols = hb_arrayLen(pFields);
+    cols = (int)hb_arrayLen(pFields);
 
     for (icol = 1; icol <= cols; icol++) {
       //        HB_LONG lType;
@@ -1303,7 +1303,7 @@ HB_FUNC(FBLINEPROCESSED)
             //                   sr_TraceLog("fb.log","valor float = %lf data %s\n",*(float ISC_FAR *)
             //                   (var->sqldata),data); FBFieldGet(hb_arrayGetItemPtr(pFields, icol), temp, (char *)
             //                   data, strlen(data), bQueryOnly, ulSystemID, bTranslate);
-            hb_itemPutNDLen(temp, *(float ISC_FAR *)(var->sqldata), lLen, lDec);
+            hb_itemPutNDLen(temp, *(float ISC_FAR *)(var->sqldata), (int)lLen, (int)lDec);
             hb_arraySetForward(pRet, icol, temp);
             break;
           }
@@ -1313,7 +1313,7 @@ HB_FUNC(FBLINEPROCESSED)
             //                   sr_TraceLog("fb.log","valor double = %lf data %s\n",*(float ISC_FAR *)
             //                   (var->sqldata),data); FBFieldGet(hb_arrayGetItemPtr(pFields, icol), temp, (char *)
             //                   data, strlen(data), bQueryOnly, ulSystemID, bTranslate);
-            hb_itemPutNDLen(temp, *(double ISC_FAR *)(var->sqldata), lLen, lDec);
+            hb_itemPutNDLen(temp, *(double ISC_FAR *)(var->sqldata), (int)lLen, (int)lDec);
             hb_arraySetForward(pRet, icol, temp);
             break;
           }
